🐛 fix: 툴박스 관련 로직 코치마크로 이동
